re arrange test filest
